Boundary Control Entity structure established
