+++ player-app/src/main/java/net/wasdev/gameon/player/control/PlayerFilter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package net.wasdev.gameon.player;
+package net.wasdev.gameon.player.control;
 
 import java.io.FileInputStream;
 import java.io.IOException;
