+++ player-app/src/main/java/net/wasdev/gameon/player/entity/Player.java
@@ -1,4 +1,4 @@
-/*******************************************************************************
+/** *****************************************************************************
  * Copyright (c) 2015 IBM Corp.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,34 +12,33 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *******************************************************************************/
-package net.wasdev.gameon.player;
-
-import java.nio.ByteBuffer;
-import java.util.Base64;
-import java.util.Base64.Encoder;
-import java.util.UUID;
-
-import org.ektorp.support.CouchDbDocument;
+ ****************************************************************************** */
+package net.wasdev.gameon.player.entity;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import java.nio.ByteBuffer;
+import java.util.Base64;
+import java.util.Base64.Encoder;
+import java.util.UUID;
+import org.ektorp.support.CouchDbDocument;
 
 @JsonInclude(Include.NON_EMPTY)
-public class Player extends CouchDbDocument{
+public class Player extends CouchDbDocument {
 
     private static final long serialVersionUID = 1L;
-   
+
     private String name;
     private String apiKey;
     private String authBy;
     private String location;
     private String favoriteColor;
-    
+
     @JsonCreator
-    public Player() {}
+    public Player() {
+    }
 
     public String getName() {
         return name;
@@ -48,7 +47,7 @@ public class Player extends CouchDbDocument{
     public void setName(String name) {
         this.name = name;
     }
-    
+
     public String getApiKey() {
         return apiKey;
     }
@@ -83,16 +82,16 @@ public class Player extends CouchDbDocument{
 
     @Override
     public String toString() {
-        return "Player [id=" + getId() + ", revision=" + getRevision() +", name=" + name + ", authBy=" + authBy + ", location=" + location
+        return "Player [id=" + getId() + ", revision=" + getRevision() + ", name=" + name + ", authBy=" + authBy + ", location=" + location
                 + ", favoriteColor=" + favoriteColor + "]";
     }
 
     //package visibility so can be invoked from places loading object
     @JsonIgnore
-    void generateApiKey(){
+    public void generateApiKey() {
         Encoder e = Base64.getEncoder();
-        ByteBuffer bb = ByteBuffer.wrap(new byte[16*2]);
-        for(int i=0;i<2; i++){
+        ByteBuffer bb = ByteBuffer.wrap(new byte[16 * 2]);
+        for (int i = 0; i < 2; i++) {
             UUID u = UUID.randomUUID();
             bb.putLong(u.getMostSignificantBits());
             bb.putLong(u.getLeastSignificantBits());
